Clarify the initial and renewal term for historical customers
Prior to May 2, 2024, the terms of service defaulted to 12 months for both the initial term and renewal term. In our updated terms of service, we defer to the Order Form, however many customers originally signed with Order Forms that did not specify a term, so the update made it unclear what term they had previously committed to.

This update makes the term length explicit.